+++ Runtime/safe-ds/safeds/data/tabular/transformation/_label_encoder.py
