+++ Runtime/safe-ds/safeds/data/tabular/transformation/_one_hot_encoder.py
