+++ Runtime/safe-ds/safeds/data/tabular/transformation/_ordinal_encoder.py
