+++ Runtime/safe-ds/tests/data/_label_encoder/test_transform.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import pytest
-from safeds.data import LabelEncoder
+from safeds.data.tabular.transformation import LabelEncoder
 from safeds.data.tabular import Table
 from safeds.exceptions import NotFittedError
