+++ Runtime/safe-ds/tests/data/_one_hot_encoder/test_fit_transform.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from safeds.data import OneHotEncoder
+from safeds.data.tabular.transformation import OneHotEncoder
 from safeds.data.tabular import Table
