+++ Runtime/safe-ds/tests/data/_one_hot_encoder/test_inverse_transform.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import pytest
-from safeds.data import OneHotEncoder
+from safeds.data.tabular.transformation import OneHotEncoder
 from safeds.data.tabular import Table
 from safeds.exceptions import NotFittedError
