+++ Runtime/safe-ds/tests/data/_ordinal_encoder/test_fit_transform.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import pytest
-from safeds.data import OrdinalEncoder
+from safeds.data.tabular.transformation import OrdinalEncoder
 from safeds.data.tabular import Table
 from safeds.data.tabular.typing import IntColumnType
 from safeds.exceptions import NotFittedError
