+++ Runtime/safe-ds/tests/data/_ordinal_encoder/test_inverse_transform.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from safeds.data import OrdinalEncoder
+from safeds.data.tabular.transformation import OrdinalEncoder
 from safeds.data.tabular import Table
 from safeds.data.tabular.typing import StringColumnType
